Changed bg, replaced UIColor extensions
