+++ MovieQuiz/Helpers/UIColor+Extensions.swift
@@ -1,4 +1,12 @@
 import Foundation
 import UIKit
 
-extension UIColor { }
+extension UIColor {
+    static let ypBackground = UIColor(named: "YP background")!
+    static let ypBlack = UIColor(named: "YP black")!
+    static let ypGreen = UIColor(named: "YP green")!
+    static let ypGrey = UIColor(named: "YP grey")!
+    static let ypRed = UIColor(named: "YP red")!
+    static let ypWhite = UIColor(named: "YP white")!
+}
+
