+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -12,7 +12,7 @@ final class MovieQuizViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        view.backgroundColor = .ypBackground
+        view.backgroundColor = .ypBlack
         let currentQuestion = questions[currentQuestionIndex]
         let viewModel = convert(model: currentQuestion)
         show(quiz: viewModel)
