--- MovieQuiz/Presentation/UIColor + Extension.swift
@@ -1,10 +1,0 @@
-import UIKit
-
-extension UIColor {
-    static let ypBackground = UIColor(named: "YP background")!
-    static let ypBlack = UIColor(named: "YP black")!
-    static let ypGreen = UIColor(named: "YP green")!
-    static let ypGrey = UIColor(named: "YP grey")!
-    static let ypRed = UIColor(named: "YP red")!
-    static let ypWhite = UIColor(named: "YP white")!
-}
